:bug: Fix: 수면 시간 단위 0.5로 수정

+++ src/main/java/com/luna/skin/domain/skin/entity/TodaySkin.java
@@ -41,7 +41,7 @@ public class TodaySkin extends BaseTimeEntity {
     private String rightImageUrl;
 
     @Column(name = "sleep_time")
-    private Integer sleepTime;
+    private Double sleepTime;
 
     @Column(name = "water_intake")
     private Double waterIntake;
